Fix native model picker visibility
