+++ apps/mobile/lib/auth-client.ts
@@ -5,7 +5,7 @@ import { createAuthClient } from '@shogo/shared-app/auth'
 import { createAuthClient as createBetterAuthClient } from 'better-auth/react'
 import { expoClient } from '@better-auth/expo/client'
 import * as SecureStore from 'expo-secure-store'
-import { API_URL } from './api'
+import { API_URL } from './api-url'
 
 function normalizeCookieHeader(cookie: string | null | undefined): string | null {
   if (!cookie) return null
